Added GetTrustees to DesConnector
*Created new endpoint
*Created Trustee case class with its companion object to support implicit Json conversion
*Tests
