+++ test/uk/gov/hmrc/trustregistration/connectors/DES/GetTrusteesSpec.scala
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2016 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.trustregistration.connectors.DES
+
+import org.mockito.Matchers
+import org.mockito.Mockito._
+import org.scalatestplus.play.{OneAppPerSuite, PlaySpec}
+import play.api.libs.json.Json
+import uk.gov.hmrc.play.http._
+import uk.gov.hmrc.trustregistration.models._
+
+import scala.concurrent.duration.Duration
+import scala.concurrent.{Await, Future}
+
+
+class GetTrusteesSpec extends PlaySpec with OneAppPerSuite with DESConnectorMocks {
+
+  "Get Trustees endpoint" must {
+    "return a GetTrusteeSuccessResponse with an empty Trustee list" when {
+      "DES returns a 200 response with an empty array" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(200, Some(Json.parse("[]")))))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe GetTrusteeSuccessResponse(List())
+      }
+    }
+
+    "return a GetTrusteeSuccessResponse with a populated Trustee list" when {
+      "DES returns a 200 response with a JSON array of Trustees" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(200, Some(Json.parse("[{\"givenName\":\"Juan\"},{\"givenName\":\"Chris\"}]")))))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe GetTrusteeSuccessResponse(List(Trustee("Juan"),Trustee("Chris")))
+      }
+    }
+
+    "return a NotFoundResponse" when {
+      "DES returns a 404 response" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(404)))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe NotFoundResponse
+      }
+    }
+
+    "return a InternalServerErrorResponse" when {
+      "DES' Json response is missing" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(200, None)))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe InternalServerErrorResponse
+      }
+      "DES returns a Json response which is invalid" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(200, Some(Json.parse("{}")))))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe InternalServerErrorResponse
+      }
+      "DES returns a 500 response" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(500)))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe InternalServerErrorResponse
+      }
+      "DES returns any unspecified error response (i.e. a 418)" in {
+        when (mockHttpGet.GET[HttpResponse](Matchers.any())(Matchers.any(),Matchers.any())).thenReturn(Future.successful(HttpResponse(418)))
+        val result = Await.result(SUT.getTrustees("1234"),Duration.Inf)
+        result mustBe InternalServerErrorResponse
+      }
+    }
+  }
+}
